Split `epoll_file.rs`
